SoilScan/requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.3          # EfficientNet, ConvNeXt, ViT via timm

# Data
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.3.0
scikit-learn>=1.4.0  # metrics, stratified splits

# Config
PyYAML>=6.0.1

# Logging & experiment tracking
tensorboard>=2.16.0

# Visualisation (optional, used in evaluate.py)
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
